+++ packages/blade-coverage-extension/background-scripts/blade-coverage-script.js
@@ -1,5 +1,12 @@
 /* eslint-disable no-undef */
 const calculateBladeCoverage = (shouldHighlightNodes) => {
+  const bladeElementExceptions = [
+    // table library adds a div internally which we want to skip
+    '[data-blade-component="table-cell"] > div', 
+    '[data-blade-component="table-header-cell"] > div', 
+    '[data-blade-component="table-footer-cell"] > div'
+  ];
+
   /**
    * Checks if DOM node is hidden or not
    */
@@ -61,6 +68,10 @@ const calculateBladeCoverage = (shouldHighlightNodes) => {
     }
     // if it's a svg node but not a blade icon then skip it
     if (closestSvgNode && !elm.hasAttribute('data-blade-component')) {
+      return;
+    }
+
+    if (bladeElementExceptions.some((exception) => elm.matches(exception))) {
       return;
     }