+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @razorpay/blade
 
+## 12.71.0
+
+### Minor Changes
+
+- b6c2e2563: feat(blade): expose tick formatter
+
+## 12.70.1
+
+### Patch Changes
+
+- 1005547fa: blade(fix): allow users to modify interval and other props
+
+## 12.70.0
+
+### Minor Changes
+
+- ea3f318d9: feat(blade): add highlight lengends on click
+- d4e6866b7: feat(blade): add support for text truncation in x-axis label
+- b1d21bf26: feat(blade): add hover on line charts
+
+## 12.69.0
+
+### Minor Changes
+
+- 9f7cbd5ad: feat(blade): bundle recharts
+
 ## 12.68.1
 
 ### Patch Changes